// frontend/src/api/model/githubCom4H1RZooraInternalDomainAttendance.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainAttendanceStatus } from "./githubCom4H1RZooraInternalDomainAttendanceStatus"
import type { GithubCom4H1RZooraInternalDomainClass } from "./githubCom4H1RZooraInternalDomainClass"
import type { GithubCom4H1RZooraInternalDomainClassSession } from "./githubCom4H1RZooraInternalDomainClassSession"
import type { GithubCom4H1RZooraInternalDomainUser } from "./githubCom4H1RZooraInternalDomainUser"

export interface GithubCom4H1RZooraInternalDomainAttendance {
  class?: GithubCom4H1RZooraInternalDomainClass
  class_id?: string
  class_session?: GithubCom4H1RZooraInternalDomainClassSession
  class_session_id?: string
  created_at?: string
  id?: string
  is_auto_marked?: boolean
  organization_id?: string
  remarks?: string
  status?: GithubCom4H1RZooraInternalDomainAttendanceStatus
  updated_at?: string
  user?: GithubCom4H1RZooraInternalDomainUser
  user_id?: string
}